+++ tests/packageParser.js
@@ -1,12 +1,16 @@
 import { resolve as resolvePath } from 'path';
+import { sep as pathSeparator } from 'path';
 import valid from './__fixtures__/packageParser/valid.json';
 import validExports from './__fixtures__/packageParser/validExports.json';
+import validSubPathExports from './__fixtures__/packageParser/validSubPathExports.json';
 import packageParser from '../src/packageParser.js';
 import { deepClone } from './__helpers__/utils';
 
 const fixturesPath = resolvePath( __dirname, '__fixtures__', 'packageParser' );
 const validFixturePath = resolvePath( fixturesPath, 'valid.json' );
 const invalidFixturePath = resolvePath( fixturesPath, 'invalid.json' );
+const invalidCJSMetdataError = 'Package metadata must contain one of "exports[ \'.\' ].require", "exports.require" or "main" properties or all of them.';
+const invalidESMMetdataError = 'Package metadata must contain one of "exports[ \'.\' ].import", "exports.import", "module" or "jsnext:main" properties or all of them.';
 
 describe( 'packageParser', () => {
 	it( 'is a function', () => {
@@ -55,7 +59,7 @@ describe( 'packageParser', () => {
 				name: 'test',
 				version: '0.0.0'
 			} );
-		} ).to.throw( ReferenceError, 'Package metadata must contain one of "exports.require" or "main" properties or all of them.' );
+		} ).to.throw( ReferenceError, invalidCJSMetdataError );
 
 		expect( () => {
 			packageParser( {
@@ -63,7 +67,7 @@ describe( 'packageParser', () => {
 				version: '0.0.0',
 				main: 'test'
 			} );
-		} ).to.throw( ReferenceError, 'Package metadata must contain one of "exports.import", "module" or "jsnext:main" properties or all of them.' );
+		} ).to.throw( ReferenceError, invalidESMMetdataError );
 
 		expect( () => {
 			packageParser( {
@@ -95,7 +99,7 @@ describe( 'packageParser', () => {
 					require: 'dist/whatever.js'
 				}
 			} );
-		} ).to.throw( ReferenceError, 'Package metadata must contain one of "exports.import", "module" or "jsnext:main" properties or all of them.' );
+		} ).to.throw( ReferenceError, invalidESMMetdataError );
 	} );
 
 	// #61
@@ -108,7 +112,7 @@ describe( 'packageParser', () => {
 					import: 'dist/whatever.js'
 				}
 			} );
-		} ).to.throw( ReferenceError, 'Package metadata must contain one of "exports.require" or "main" properties or all of them.' );
+		} ).to.throw( ReferenceError, invalidCJSMetdataError );
 	} );
 
 	it( 'returns simplified metadata', () => {
@@ -117,10 +121,11 @@ describe( 'packageParser', () => {
 			author: 'Comandeer',
 			license: 'MIT',
 			version: '9.0.1',
-			src: 'src/index.js',
 			dist: {
-				esm: 'dist/es2015.js',
-				cjs: 'dist/es5.js'
+				[ `src${ pathSeparator }index.js` ]: {
+					esm: 'dist/es2015.js',
+					cjs: 'dist/es5.js'
+				}
 			}
 		} );
 	} );
@@ -132,35 +137,83 @@ describe( 'packageParser', () => {
 			author: 'Comandeer',
 			license: 'MIT',
 			version: '9.0.1',
-			src: 'src/index.js',
 			dist: {
-				esm: 'dist/test-package.mjs',
-				cjs: 'dist/test-package.cjs'
+				[ `src${ pathSeparator }index.js` ]: {
+					esm: 'dist/test-package.mjs',
+					cjs: 'dist/test-package.cjs'
+				}
 			}
 		} );
+	} );
+
+	// #185
+	it( 'returns simplified metadata for package with subpath "exports" field', () => {
+		expect( packageParser( validSubPathExports ) ).to.deep.equal( {
+			name: 'test-package',
+			author: 'Comandeer',
+			license: 'ISC',
+			version: '1.0.0',
+			dist: {
+				[ `src${ pathSeparator }index.js` ]: {
+					cjs: './dist/es5.cjs',
+					esm: './dist/es6.mjs'
+				},
+				[ `src${ pathSeparator }chunk.js` ]: {
+					cjs: './dist/not-related-name.cjs',
+					esm: './dist/also-not-related-name.js'
+				}
+			}
+		} );
+	} );
+
+	// #185
+	it( 'prefers exports[ \'.\' ].import over exports.import', () => {
+		const distPath = 'dist/subpath.mjs';
+		const module = deepClone( validExports );
+		module.exports[ '.' ] = {
+			import: distPath
+		};
+		const indexDistMetadata = parseMetadataAndGetDistInfo( module );
+
+		expect( indexDistMetadata.esm ).to.equal( distPath );
+	} );
+
+	// #185
+	it( 'prefers exports[ \'.\' ].require over exports.require', () => {
+		const distPath = `dist${ pathSeparator }subpath.cjs`;
+		const module = deepClone( validExports );
+		module.exports[ '.' ] = {
+			require: distPath
+		};
+		const indexDistMetadata = parseMetadataAndGetDistInfo( module );
+
+		expect( indexDistMetadata.cjs ).to.equal( distPath );
 	} );
 
 	// #61
 	it( 'prefers exports.import over module', () => {
 		const module = deepClone( validExports );
 		module.module = 'dist/es2015.js';
+		const indexDistMetadata = parseMetadataAndGetDistInfo( module );
 
-		expect( packageParser( module ).dist.esm ).to.equal( 'dist/test-package.mjs' );
+		expect( indexDistMetadata.esm ).to.equal( 'dist/test-package.mjs' );
 	} );
 
 	// #61
 	it( 'prefers exports.require over main', () => {
 		const module = deepClone( validExports );
 		module.main = 'dist/es5.js';
+		const indexDistMetadata = parseMetadataAndGetDistInfo( module );
 
-		expect( packageParser( module ).dist.cjs ).to.equal( 'dist/test-package.cjs' );
+		expect( indexDistMetadata.cjs ).to.equal( 'dist/test-package.cjs' );
 	} );
 
 	it( 'prefers module over jsnext:main', () => {
 		const module = deepClone( valid );
 		module[ 'jsnext:main' ] = 'dist/esnext.js';
+		const indexDistMetadata = parseMetadataAndGetDistInfo( module );
 
-		expect( packageParser( module ).dist.esm ).to.equal( 'dist/es2015.js' );
+		expect( indexDistMetadata.esm ).to.equal( 'dist/es2015.js' );
 	} );
 
 	// #61
@@ -168,8 +221,9 @@ describe( 'packageParser', () => {
 		const module = deepClone( validExports );
 		module.module = 'dist/module.js';
 		delete module.exports.import;
+		const indexDistMetadata = parseMetadataAndGetDistInfo( module );
 
-		expect( packageParser( module ).dist.esm ).to.equal( 'dist/module.js' );
+		expect( indexDistMetadata.esm ).to.equal( 'dist/module.js' );
 	} );
 
 	// #61
@@ -177,8 +231,9 @@ describe( 'packageParser', () => {
 		const module = deepClone( validExports );
 		module[ 'jsnext:main' ] = 'dist/jsnext.js';
 		delete module.exports.import;
+		const indexDistMetadata = parseMetadataAndGetDistInfo( module );
 
-		expect( packageParser( module ).dist.esm ).to.equal( 'dist/jsnext.js' );
+		expect( indexDistMetadata.esm ).to.equal( 'dist/jsnext.js' );
 	} );
 
 	// #61
@@ -186,16 +241,18 @@ describe( 'packageParser', () => {
 		const module = deepClone( validExports );
 		module.main = 'dist/legacy.js';
 		delete module.exports.require;
+		const indexDistMetadata = parseMetadataAndGetDistInfo( module );
 
-		expect( packageParser( module ).dist.cjs ).to.equal( 'dist/legacy.js' );
+		expect( indexDistMetadata.cjs ).to.equal( 'dist/legacy.js' );
 	} );
 
 	it( 'uses jsnext:main when module is not available', () => {
-		const jsnext = deepClone( valid );
-		jsnext[ 'jsnext:main' ] = 'dist/esnext.js';
-		delete jsnext.module;
+		const module = deepClone( valid );
+		module[ 'jsnext:main' ] = 'dist/esnext.js';
+		delete module.module;
+		const indexDistMetadata = parseMetadataAndGetDistInfo( module );
 
-		expect( packageParser( jsnext ).dist.esm ).to.equal( 'dist/esnext.js' );
+		expect( indexDistMetadata.esm ).to.equal( 'dist/esnext.js' );
 	} );
 
 	it( 'parses author object into string', () => {
@@ -207,3 +264,9 @@ describe( 'packageParser', () => {
 		expect( packageParser( author ).author ).to.equal( 'Tester' );
 	} );
 } );
+
+function parseMetadataAndGetDistInfo( metadata, srcFile = `src${ pathSeparator }index.js` ) {
+	const parsedMetadata = packageParser( metadata );
+
+	return parsedMetadata.dist[ srcFile ];
+}