+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ---
 
 ## [0.15.0]
+### Added
+* [#185]: Support for subpath exports.
 ### Changed
 * [#199]: **BREAKING CHANGE**: update dependencies including major versions of:
 	* `@rollup/plugin-commonjs` from 19.x to `^21.0.0`.
@@ -173,6 +175,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [#173]: https://github.com/Comandeer/rollup-lib-bundler/issues/173
 [#176]: https://github.com/Comandeer/rollup-lib-bundler/issues/176
 [#179]: https://github.com/Comandeer/rollup-lib-bundler/issues/179
+[#185]: https://github.com/Comandeer/rollup-lib-bundler/issues/185
 [#189]: https://github.com/Comandeer/rollup-lib-bundler/issues/189
 [#191]: https://github.com/Comandeer/rollup-lib-bundler/issues/191
 [#193]: https://github.com/Comandeer/rollup-lib-bundler/issues/193